Moved the pages to pages folder
