+++ src/pages/StartPage/StartPage.css
